Cleanup: use consistent folder and naming structure in examples
