+++ example/counter/counter.hpp
@@ -12,6 +12,7 @@
 
 #include <lager/debug/cereal/struct.hpp>
 #include <lager/util.hpp>
+
 #include <variant>
 
 namespace counter {
@@ -34,13 +35,12 @@ using action = std::variant<increment_action, decrement_action, reset_action>;
 
 inline model update(model c, action action)
 {
-    return std::visit(
-        lager::visitor{
-            [&](increment_action) { return model{c.value + 1}; },
-            [&](decrement_action) { return model{c.value - 1}; },
-            [&](reset_action a) { return model{a.new_value}; },
-        },
-        action);
+    return std::visit(lager::visitor{
+                          [&](increment_action) { return model{c.value + 1}; },
+                          [&](decrement_action) { return model{c.value - 1}; },
+                          [&](reset_action a) { return model{a.new_value}; },
+                      },
+                      action);
 }
 
 LAGER_CEREAL_STRUCT(model, (value));
