+++ example/todo/imgui/main.cpp
@@ -1,0 +1,94 @@
+//
+// lager - library for functional interactive c++ programs
+// Copyright (C) 2017 Juan Pedro Bolivar Puente
+//
+// This file is part of lager.
+//
+// lager is free software: you can redistribute it and/or modify
+// it under the terms of the MIT License, as detailed in the LICENSE
+// file located at the root of this source code distribution,
+// or here: <https://github.com/arximboldi/lager/blob/master/LICENSE>
+//
+
+#include "../todo.hpp"
+
+#include <lager/event_loop/sdl.hpp>
+#include <lager/store.hpp>
+
+#include <imgui.h>
+#include <imgui_impl_opengl2.h>
+#include <imgui_impl_sdl.h>
+
+#include <SDL.h>
+#include <SDL_opengl.h>
+
+void draw(const todo::model& m) {}
+
+int main()
+{
+    SDL_Init(SDL_INIT_VIDEO | SDL_INIT_TIMER | SDL_INIT_GAMECONTROLLER);
+
+    SDL_GL_SetAttribute(SDL_GL_DOUBLEBUFFER, 1);
+    SDL_GL_SetAttribute(SDL_GL_DEPTH_SIZE, 24);
+    SDL_GL_SetAttribute(SDL_GL_STENCIL_SIZE, 8);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MAJOR_VERSION, 2);
+    SDL_GL_SetAttribute(SDL_GL_CONTEXT_MINOR_VERSION, 2);
+
+    auto window =
+        SDL_CreateWindow("Todo Imgui",
+                         SDL_WINDOWPOS_CENTERED,
+                         SDL_WINDOWPOS_CENTERED,
+                         1280,
+                         720,
+                         SDL_WINDOW_OPENGL | SDL_WINDOW_RESIZABLE |
+                             SDL_WINDOW_ALLOW_HIGHDPI | SDL_WINDOW_SHOWN);
+    auto gl_context = SDL_GL_CreateContext(window);
+    SDL_GL_MakeCurrent(window, gl_context);
+    SDL_GL_SetSwapInterval(1); // Enable vsync
+
+    IMGUI_CHECKVERSION();
+    ImGui::CreateContext();
+    auto& io = ImGui::GetIO();
+    io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard;
+    io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;
+
+    ImGui::StyleColorsDark();
+
+    ImGui_ImplSDL2_InitForOpenGL(window, gl_context);
+    ImGui_ImplOpenGL2_Init();
+
+    auto loop  = lager::sdl_event_loop{};
+    auto store = lager::make_store<todo::action>(
+        todo::model{}, todo::update, lager::with_sdl_event_loop{loop});
+
+    lager::watch(store, [window](auto&&, auto&& m) {
+        ImGui_ImplOpenGL2_NewFrame();
+        ImGui_ImplSDL2_NewFrame(window);
+        ImGui::NewFrame();
+
+        draw(m);
+
+        ImGui::Render();
+        auto size = ImGui::GetIO().DisplaySize;
+        glViewport(0, 0, (int) size.x, (int) size.y);
+        glClearColor(0, 0, 0, 1);
+        glClear(GL_COLOR_BUFFER_BIT);
+        ImGui_ImplOpenGL2_RenderDrawData(ImGui::GetDrawData());
+        SDL_GL_SwapWindow(window);
+    });
+
+    loop.run([&](const SDL_Event& ev) {
+        ImGui_ImplSDL2_ProcessEvent(&ev);
+        return ev.type == SDL_QUIT;
+    });
+
+    ImGui_ImplOpenGL2_Shutdown();
+    ImGui_ImplSDL2_Shutdown();
+    ImGui::DestroyContext();
+
+    SDL_GL_DeleteContext(gl_context);
+    SDL_DestroyWindow(window);
+    SDL_Quit();
+
+    return 0;
+}
